add more sems

+++ Kernel/include/sync.h
@@ -4,12 +4,12 @@
 #include <stddef.h>
 #include <stdint.h>
 
-int sem_open(uint8_t id, uint64_t value);
+int sem_open(uint16_t id, uint64_t value);
 
-int sem_wait(uint8_t semID);
+int sem_wait(uint16_t semID);
 
-int sem_post(uint8_t semID);
+int sem_post(uint16_t semID);
 
-int sem_close(uint8_t semID);
+int sem_close(uint16_t semID);
 
 #endif
